Big bugged (probably) push for whole bunch of things

+++ install.sh
@@ -1,33 +1,31 @@
 . ./readline/rlwrap_scripts.sh
 . ./checks/check_distro.sh
+. ./checks/check_rlwrap.sh
 
-reade -Q "GREEN" -P "y" -p "Install moar? (Pager with linenumbers) [Y/n]: " "y n" moar
-if [ -z $moar ] || [ "Y" == $moar ] || [ $moar == "y" ]; then
-   . ./install_moar.sh 
+
+if [ ! -e ~/config ]; then
+    reade -Q "BLUE" -i "y" -p "Create ~/.config to ~/config symlink? [Y/n]:" "y n" sym1
+    if [ -z $sym1 ] || [ "y" == $sym1 ] && [ ! -e ~/config ]; then
+        ln -s ~/.config ~/config
+    fi
 fi
 
-reade -Q "GREEN" -P "y" -p "Install fzf? (Fuzzy file/folder finder (better reverse-search)) [Y/n]: " "y n" findr
-if [ -z $findr ] || [ "Y" == $findr ] || [ $findr == "y" ]; then
-   . ./install_fzf.sh 
+if [ ! -e ~/lib_systemd ]; then
+    reade -Q "BLUE" -i "y" -p "Create /lib/systemd/system/ to user directory symlink? [Y/n]:" "y n" sym2
+    if [ -z $sym2 ] || [ "y" == $sym2 ] && [ ! -e ~/lib_systemd ]; then
+        ln -s /lib/systemd/system/ ~/lib_systemd
+    fi
 fi
 
-read -p "Create ~/.config to ~/config symlink? [Y/n]:" sym1
-if [ -z $sym1 ] || [ "y" == $sym1 ] && [ ! -e ~/config ]; then
-    ln -s ~/.config ~/config
-fi
-
-read -p "Create /lib/systemd/system/ to user directory symlink? [Y/n]:" sym2
-if [ -z $sym2 ] || [ "y" == $sym2 ] && [ ! -e ~/lib_systemd ]; then
-    ln -s /lib/systemd/system/ ~/lib_systemd
-fi
-
-read -p "Create /etc/systemd/system/ to user directory symlink? [Y/n]:" sym3
-if [ -z $sym3 ] || [ "y" == $sym3 ] && [ ! -e ~/etc_systemd ]; then
-    ln -s /etc/systemd/system/ ~/etc_systemd
+if [ ! -e ~/etc_systemd ]; then
+    reade -Q "BLUE" -i "y" -p "Create /etc/systemd/system/ to user directory symlink? [Y/n]:" "y n" sym3
+    if [ -z $sym3 ] || [ "y" == $sym3 ] && [ ! -e ~/etc_systemd ]; then
+        ln -s /etc/systemd/system/ ~/etc_systemd
+    fi
 fi
 
 
-reade -Q "GREEN" -P "y" -p "Create ~/.bash_aliases.d/, link it to .bashrc and install scripts? (Readline included) [Y/n]:" "y n" scripts
+reade -Q "GREEN" -i "y" -p "Check existence (and create) ~/.bash_aliases.d/, link it to .bashrc and start installation scripts? [Y/n]:" "y n" scripts
 if [ -z $scripts ] || [ "y" == $scripts ]; then
 
     if [ ! -d ~/.bash_aliases.d/ ]; then
@@ -48,7 +46,7 @@ if [ -z $scripts ] || [ "y" == $scripts ]; then
     cp -fv readline/rlwrap_scripts.sh ~/.bash_aliases.d/rlwrap_scripts.sh
 
 
-    reade -Q "YELLOW" -P "y" -p "Create /root/.bash_aliases.d/, link it to /root/.bashrc [Y/n]:" "y n" rscripts
+    reade -Q "YELLOW" -i "y" -p "Create /root/.bash_aliases.d/, link it to /root/.bashrc [Y/n]:" "y n" rscripts
     if [ -z $rscripts ] || [ "y" == $rscripts ]; then
 
         if ! sudo test -d /root/.bash_aliases.d/ ; then
@@ -64,30 +62,117 @@ if [ -z $scripts ] || [ "y" == $scripts ]; then
         sudo cp -fv readline/rlwrap_scripts.sh ~/.bash_aliases.d/rlwrap_scripts.sh 
     fi
     
-    function inputrc_r(){
-         sudo cp -fv readline/.inputrc /root/.inputrc
-         sudo cp -fv readline/shell_keybindings.sh /root/.bash_aliases.d/
-    }
-    function inputrc() {
-        cp -fv readline/.inputrc ~/
-        cp -fv readline/shell_keybindings.sh ~/.bash_aliases.d/
-        yes_edit_no inputrc_r "readline/.inputrc readline/shell_keybindings.sh" "Install .inputrc at /root/?" "edit" "RED"
-    }
-    yes_edit_no inputrc "readline/.inputrc readline/shell_keybindings.sh" "Install .inputrc at ~/ ? (readline config)" "edit" "YELLOW"
 
-
+    # Xresources
+    
     function xresources_r(){
         sudo cp -fv xterm/.Xresources /root/.Xresources
+        sudo cp -fv xterm/xterm.sh /root/xterm.sh
     }
     function xresources(){
         cp -fv xterm/.Xresources ~/.Xresources
-        yes_edit_no xresources_r "xterm/.Xresources" "Install .Xresources at /root/?" "edit" "RED"
+        cp -fv xterm/xterm.sh ~/xterm.sh
+        yes_edit_no xresources_r "xterm/.Xresources xterm/xterm.sh" "Install .Xresources and xterm.sh at /root/?" "edit" "RED"
     }
-    yes_edit_no xresources "xterm/.Xresources" "Install .Xresources at ~/ ? (readline config)" "edit" "YELLOW"
+    yes_edit_no xresources "xterm/.Xresources xterm/xterm.sh" "Install .Xresources and xterm.sh at ~/.bash_aliases.d/ ? (readline config)" "edit" "YELLOW"
+    
+    # Tty keybinding
 
+    stty_r(){
+         sudo cp -fv aliases/01-tty.sh /root/.bash_aliases.d/
+    }
+    stty_n(){
+        cp -fv aliases/01-tty.sh ~/.bash_aliases.d/
+        yes_edit_no stty_r "aliases/01-tty.sh" "Install 01-stty.sh at /root/.bash_aliases.d/?" "yes" "YELLOW"
+    }
+    yes_edit_no stty_n "aliases/01-tty.sh" "Install 01-stty.sh at ~/.bash_aliases.d/ (Tty keybindings and other startup lines) ? " "edit" "YELLOW"
 
+    # Readline
+    
+     function inputrc_r(){
+         sudo cp -fv readline/.inputrc /root/.inputrc
+    }
+    function inputrc() {
+        cp -fv readline/.inputrc ~/
+        yes_edit_no inputrc_r "readline/.inputrc" "Install .inputrc at /root/?" "edit" "GREEN"
+    }
+    yes_edit_no inputrc "readline/.inputrc" "Install .inputrc at ~/ ? (readline config)" "edit" "GREEN"
+    
+    # Shell-keybindings
+    
+    if [ grep -q "bind -x '\"\C-s\": ctrl-s'" ./aliases/shell_keybindings.sh ] && [ ! grep -q "#bind -x '\"\C-s\": ctrl-s'" ./aliases/shell_keybindings.sh ]; then
+        sed -i 's|bind -x '\''"\\C-s\": ctrl-s'\''|#bind -x '\''"\\C-s\": ctrl-s'\''|g' ./aliases/shell_keybindings.sh
+        sed -i 's|bind -x '\''"\\eOR\": ctrl-s'\''|#bind -x '\''"\\eOR\": ctrl-s'\''|g' ./aliases/shell_keybindings.sh
+    fi
 
-    reade -Q "GREEN" -P "y" -p "Install bash completions for aliases in ~/.bash_completion.d? [Y/n]:" "y n" compl
+    if grep -q "bind -x '\"\eOQ\": ranger'" readline/shell_keybindings.sh; then
+        sed -i 's|bind -x '\''"\\eOQ\": ranger'\''|#bind -x '\''"\\eOQ\": ranger'\''|g' ./readline/shell_keybindings.sh
+    fi
+
+    function shell-keybindings_r(){
+         sudo cp -fv aliases/shell_keybindings.sh /root/.bash_aliases.d/
+    }
+    function shell-keybindings() {
+        cp -fv aliases/shell_keybindings.sh ~/.bash_aliases.d/
+        yes_edit_no shell-keybindings_r "aliases/shell_keybindings.sh" "Install .bash_aliases.d/shell_keybindings.sh at /root/?" "edit" "RED"
+    }
+    yes_edit_no shell-keybindings_r "aliases/shell_keybindings.sh" "Install .bash_aliases.d/shell-keybindings.sh at ~/ ? (bind commands)" "edit" "YELLOW"
+
+    # Pathvariables
+    if [ grep -q "export MOAR=" aliases/00-pathvariables.sh ] && [ ! grep -q "#export MOAR=" aliases/00-pathvariables.sh ]; then
+        sed -i "s/export MOAR=/#export MOAR=/g" aliases/00-pathvariables.sh 
+        sed -i "s/export PAGER=/#export PAGER=/g" aliases/00-pathvariables.sh
+        sed -i "s/export SYSTEMD_PAGER=/#export SYSTEMD_PAGER=/g" aliases/00-pathvariables.sh
+        sed -i "s/export SYSTEMD_PAGERSECURE=/#export SYSTEMD_PAGERSECURE=/g" aliases/00-pathvariables.sh
+    fi
+
+    if [ grep -q "export FZF_DEFAULT_COMMAND" aliases/00-pathvariables.sh ] && [ ! grep -q "#export FZF_DEFAULT_COMMAND" aliases/00-pathvariables.sh ]; then
+        sed -i 's|export FZF_DEFAULT_COMMAND|#export FZF_DEFAULT_COMMAND|g' aliases/00-pathvariables.sh
+        sed -i 's|export FZF_CTRL_T_COMMAND|#export FZF_CTRL_T_COMMAND|g' aliases/00-pathvariables.sh  
+    fi
+
+    pathvariables_r(){
+         sudo cp -fv aliases/00-pathvariables.sh /root/.bash_aliases.d/00-pathvariables.sh
+    }                                            
+    pathvariables(){
+        cp -fv aliases/00-pathvariables.sh ~/.bash_aliases.d/00-pathvariables.sh
+        yes_edit_no pathvariables_r "aliases/00-pathvariables.sh" "Install 00-pathvariables.sh at /root/.bash_aliases.d/?" "yes" "YELLOW"
+    }
+    yes_edit_no pathvariables "aliases/00-pathvariables.sh" "Install 00-pathvariables.sh at ~/.bash_aliases.d/ ? " "edit" "YELLOW"
+    
+    # Moar (Custom pager instead of less)
+    
+    if [ ! -x "$(command -v moar)" ]; then
+        reade -Q "GREEN" -i "y" -p "Install moar? (Custom pager instead of less with linenumbers) [Y/n]: " "y n" moar
+        if [ -z $moar ] || [ "Y" == $moar ] || [ $moar == "y" ]; then
+            . ./install_moar.sh 
+            reade -Q "GREEN" -i "y" -p "Set moar default pager for $USER? [Y/n]: " "y n" moar_usr
+            if [ -z "$moar_usr" ] || [ "y" == "$moar_usr" ] || [ "Y" == "$moar_usr" ]; then
+                sed -i "s/#export MOAR=/export MOAR=/g" ~/aliases/00-pathvariables.sh 
+                sed -i "s/#export PAGER=/export PAGER=/g" ~/aliases/00-pathvariables.sh
+                sed -i "s/#export SYSTEMD_PAGER=/export SYSTEMD_PAGER=/g" ~/aliases/00-pathvariables.sh
+                sed -i "s/#export SYSTEMD_PAGERSECURE=/export SYSTEMD_PAGERSECURE=/g" ~/aliases/00-pathvariables.sh
+
+            fi
+            
+            reade -Q "YELLOW" -i "y" -p "Set moar default pager for root? [Y/n]: " "y n" moar_root
+            if [ -z "$moar_root" ] || [ "y" == "$moar_root" ] || [ "Y" == "$moar_root" ]; then
+                sudo sed -i "s/#export MOAR=/export MOAR=/g" /root/.aliases/00-pathvariables.sh 
+                sudo sed -i "s/#export PAGER=/export PAGER=/g" /root/aliases/00-pathvariables.sh
+                sudo sed -i "s/#export SYSTEMD_PAGER=/export SYSTEMD_PAGER=/g" /root/aliases/00-pathvariables.sh
+                sudo sed -i "s/#export SYSTEMD_PAGERSECURE=/export SYSTEMD_PAGERSECURE=/g" /root/aliases/00-pathvariables.sh
+
+            fi
+        fi
+    fi
+
+    # Fzf (Fuzzy Finder)
+    . ./install_fzf.sh 
+    
+    # Ranger (File explorer)
+    . ./install_ranger.sh 
+    
+    reade -Q "GREEN" -i "y" -p "Install bash completions for aliases in ~/.bash_completion.d? [Y/n]:" "y n" compl
     if [ -z $compl ] || [ "y" == $compl ]; then
         if [ ! -d ~/.bash_completion.d/ ]; then 
             mkdir ~/.bash_completion.d
@@ -100,7 +185,7 @@ if [ -z $scripts ] || [ "y" == $scripts ]; then
             echo ". ~/.bash_completion.d/complete_alias" >> ~/.bashrc
         fi
         
-        reade -Q "YELLOW" -P "y" -p "Install bash completions for aliases in /root/.bash_completion.d? [y/n]:" "y n" rcompl
+        reade -Q "YELLOW" -i "y" -p "Install bash completions for aliases in /root/.bash_completion.d? [Y/n]:" "y n" rcompl
         if [ -z $rcompl ] || [ "y" == $rcompl ]; then
             if ! sudo test -d /root/.bash_completion.d/ ; then 
                 sudo mkdir /root/.bash_completion.d
@@ -115,21 +200,11 @@ if [ -z $scripts ] || [ "y" == $scripts ]; then
         fi
     fi
     
-    reade -Q "GREEN" -P "y" -p "Install python completions in */.bash_completion.d? [Y/n]:" "y n" pycomp
-    if [ -z $pycomp ] || [ "y" == $pycomp ]; then
-        . install_pythonCompletions_bash.sh
-    fi
+    #reade -Q "GREEN" -i "y" -p "Install python completions in */.bash_completion.d? [Y/n]:" "y n" pycomp
+    #if [ -z $pycomp ] || [ "y" == $pycomp ]; then
+    #    . install_pythonCompletions_bash.sh
+    #fi
 
-    function bash_yes_r(){
-        sudo cp -fv aliases/bash.sh /root/.bash_aliases.d/;
-    }
-
-    function bash_yes() {
-        cp -fv aliases/bash.sh ~/.bash_aliases.d/;
-        yes_edit_no bash_yes_r "aliases/bash.sh" "Install bash.sh at /root/?" "yes" "YELLOW";
-    }
-
-    yes_edit_no bash_yes "aliases/bash.sh" "Install bash.sh at ~/ ? (bash specific aliases)?" "yes" "GREEN";
     
     #keybinds_r(){
         #sudo cp -fv aliases/shell_keybindings.sh /root/.bash_aliases.d
@@ -144,15 +219,16 @@ if [ -z $scripts ] || [ "y" == $scripts ]; then
     #}
     #yes_edit_no keybinds "aliases/shell_keybindings.sh" "Install shell_keybindings.sh at ~/ ? (generally related to keybinds)? " "yes" "GREEN";
     
+    function bash_yes_r(){
+        sudo cp -fv aliases/bash.sh /root/.bash_aliases.d/;
+    }
 
-    pathvariables_r(){
-         sudo cp -fv aliases/pathvariables.sh /root/.bash_aliases.d/
+    function bash_yes() {
+        cp -fv aliases/bash.sh ~/.bash_aliases.d/;
+        yes_edit_no bash_yes_r "aliases/bash.sh" "Install bash.sh at /root/?" "yes" "YELLOW";
     }
-    pathvariables(){
-        cp -fv aliases/pathvariables.sh ~/.bash_aliases.d/
-        yes_edit_no pathvariables_r "aliases/pathvariables.sh" "Install pathvariables.sh at /root/?" "yes" "YELLOW"
-    }
-    yes_edit_no pathvariables "aliases/pathvariables.sh" "Install pathvariables.sh at ~/ ? " "edit" "YELLOW"
+
+    yes_edit_no bash_yes "aliases/bash.sh" "Install bash.sh at ~/ ? (bash specific aliases)?" "yes" "GREEN";
 
 
     general_r(){
@@ -214,10 +290,12 @@ if [ -z $scripts ] || [ "y" == $scripts ]; then
 
     gitt(){
         if [ $distro == "Arch" ] || [ $distro_base == "Arch" ]; then
-            sudo pacman -Su git
+            yes | sudo pacman -Su git
         elif [ $distro == "Debian" ] || [ $distro_base == "Debian" ]; then
-            sudo apt install git
+            yes | sudo apt update
+            yes | sudo apt install git
         fi
+        #if [ git config --list | grep -q user.email ]; then
         read -p "Configure global user and email git? [Y/n]: " gitcnf
         if [ -z $gitcnf ] || [ "y" == $gitcnf ]; then
             read -p "Email: " mail
@@ -226,7 +304,9 @@ if [ -z $scripts ] || [ "y" == $scripts ]; then
             git config --global user.name "$name"
         fi 
         cp -fv aliases/git.sh ~/.bash_aliases.d/
-        yes_edit_no git_r "aliases/git.sh" "Install git.sh at /root/?" "no" "RED" 
+        
+        . install_copy-conf.sh
+        #yes_edit_no git_r "aliases/git.sh" "Install git.sh at /root/?" "no" "RED" 
     }
     yes_edit_no gitt "aliases/git.sh" "Install git.sh at ~/.bash_aliases.d/ (git aliases)? " "yes" "GREEN"
